Package box: rebase if image has backing image

+++ bin/package.sh
@@ -2,7 +2,7 @@
 
 error() {
     local msg="${1}"
-    echo "${msg}"
+    echo "==> ${msg}"
     exit 1
 }
 
@@ -11,6 +11,32 @@ usage() {
     echo
     echo "Package a kvm qcow2 image into a kvm vagrant reusable box"
     echo "It uses virt-install to do so"
+}
+
+# Print the image's backing file
+backing(){
+    local img=${1}
+    qemu-img info $img | grep 'backing file:' | cut -d ':' -f2
+}
+
+# dump libvirt xml
+dumpxml(){
+    virt-install \
+        --print-xml \
+        --dry-run \
+        --import \
+        --name $NAME \
+        --ram $RAM --vcpus=$VCPUS \
+        --disk path="$TMP_IMG",bus=virtio,format=qcow2 \
+        -w network=default,model=virtio
+    [[ "$?" -ne 0 ]] && error "Error during virt-install"
+}
+
+# Rebase the image
+rebase(){
+    local img=${1}
+    qemu-img rebase -p -b "" $img
+    [[ "$?" -ne 0 ]] && error "Error during rebase"
 }
 
 if [ -z "$2" ]; then
@@ -24,33 +50,35 @@ RAM=2048
 VCPUS=2
 
 IMG=$(readlink -e $2)
+[[ "$?" -ne 0 ]] && error "'$2': No such image"
+
 IMG_BASENAME=$(basename $IMG)
 IMG_DIR=$(dirname $IMG)
 
-# Create stuff in tmp dir
+BOX=$IMG_DIR/$NAME.box
+[[ -f "$BOX" ]] && error "'$BOX': Already exists"
+
 TMP_DIR=$IMG_DIR/_tmp_package
+TMP_IMG=$TMP_DIR/$IMG_BASENAME
+
 mkdir -p $TMP_DIR
 
-# We move the image to the tempdir
-# ensure that it's moved back again
-# and the tmp dir removed
-trap "mv $TMP_DIR/$IMG_BASENAME $IMG_DIR; rm -rf $TMP_DIR" EXIT
-
-mv $IMG $TMP_DIR
-IMG=$TMP_DIR/$IMG_BASENAME
+# We move / copy the image to the tempdir
+# ensure that it's moved back / removed again
+if [[ -n $(backing $IMG) ]]; then
+    echo "==> Image has backing image, copying image and rebasing ..."
+    trap "rm -rf $TMP_DIR" EXIT
+    cp $IMG $TMP_DIR
+    rebase $TMP_IMG
+else
+    trap "mv $TMP_DIR/$IMG_BASENAME $IMG_DIR; rm -rf $TMP_DIR" EXIT
+    mv $IMG $TMP_DIR
+fi
 
 cd $TMP_DIR
 
 # generate box.xml
-
-virt-install \
-    --print-xml \
-    --dry-run \
-    --import \
-    --name $NAME \
-    --ram $RAM --vcpus=$VCPUS \
-    --disk path="$IMG",bus=virtio,format=qcow2 \
-    -w network=default,model=virtio > box.xml
+dumpxml > box.xml
 
 # extract the mac for the Vagrantfile
 MAC=$(cat box.xml | grep 'mac address' | cut -d\' -f2 | tr -d :)
@@ -77,7 +105,11 @@ Vagrant.configure("2") do |config|
 end
 EOF
 
+echo "==> Creating box, tarring and gzipping"
+
 tar cvzf $NAME.box --totals ./metadata.json ./Vagrantfile ./box.xml ./$IMG_BASENAME
 mv $NAME.box $IMG_DIR
 
-echo "$IMG_DIR/$NAME.box created"
+echo "==> ${IMG_DIR}/${NAME}.box created"
+echo "==> You can now add the box:"
+echo "==>   'vagrant box add ${IMG_DIR}/${NAME}.box --name ${NAME%.*}'"
